+++ src/firebolt/model/V2/database.py
@@ -6,12 +6,12 @@ from datetime import datetime
 from typing import TYPE_CHECKING, ClassVar, List
 
 from firebolt.model.V2 import FireboltBaseModel
-from firebolt.service.types import EngineStatus
+from firebolt.service.V2.types import EngineStatus
 from firebolt.utils.exception import AttachedEngineInUseError
 
 if TYPE_CHECKING:
     from firebolt.model.V2.engine import Engine
-    from firebolt.service.database import DatabaseService
+    from firebolt.service.V2.database import DatabaseService
 
 logger = logging.getLogger(__name__)
