+++ src/firebolt/service/V1/base.py
