+++ src/firebolt/service/V2/instance_type.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from firebolt.model.V2.instance_type import InstanceType
-from firebolt.service.base import BaseService
+from firebolt.service.V2.base import BaseService
 from firebolt.utils.exception import InstanceTypeNotFoundError
 from firebolt.utils.urls import ACCOUNT_INSTANCE_TYPES_URL
 from firebolt.utils.util import cached_property
